move post component to utils and add its style Relates #7
