+++ client/src/components/utils/post/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import { Icon } from 'antd';
+import './style.css';
 
-export default function Post({ post }) {
+export default function Post({ postType }) {
   return (
     <button className='posts--button posts--button-white'>
       <Icon className='posts--icon posts--form-icon' type='form' />
-      {post}
+      {postType}
     </button>
   )
 }
