+++ client/src/components/utils/post/style.css
@@ -1,0 +1,17 @@
+.posts--button-white {
+  width: 100%;
+  background-color: #ffffff;
+  margin-bottom: 35px;
+  border: 1px solid rgba(144, 144, 144, .7);
+  padding: 20px 10px;
+}
+
+.posts--button-white:hover {
+  text-decoration: underline;
+}
+
+.posts--form-icon {
+  color: #D65050!important;
+  font-size: 28px;
+  padding-right: 13px;
+}
